update typehint to float

+++ bigc/api.py
@@ -4,7 +4,7 @@ from typing import Optional
 
 
 class BigCommerceAPI:
-    def __init__(self, store_hash: str, access_token: str, timeout: Optional[int] = None):
+    def __init__(self, store_hash: str, access_token: str, timeout: Optional[float] = None):
         api_client = BigCommerceAPIClient(store_hash, access_token, timeout)
 
         self.carts: BigCommerceCartsAPI = BigCommerceCartsAPI(api_client)
